Rename to `eval_error.py`.
